Okay.. the mb_ultimate target builds now. Time to do some testing on real hardware again. The most challenging will probably be the dma loads in combination with mount. Some tricky stuff is done there for finding back the mount point. Also, what is likely not going well yet, is the invalidation procedure and the file browser updates when devices are removed or changes are made to the file system.
Documentation updated: there is a new command: DOS_FILE_STAT, that allows a file to be probed and the info retrieved without opening the file.

TODO: remove the copy path from UI DOS command. This is no longer a valid thing to do, as there may be more instances of a UI.