Add options to preload a REU image on startup
Adds three new configuration options to "C64 and cartridge settings":

REU Preload (Enable/Disable)
REU Preload Image (path)
REU PRELOAD Offset (128k - 16M)

If preload is enabled, the specified image will be loaded as soon as
the corresponding filesystem is available, starting at the specified
offset. If the image size exceeds the available REU memory space, the
image is only loaded up to the end of REU memory. No action is taken
if the specified offset is greater or equal to the available REU size.

In addition, this commit adds a "Set as Preload Image" option to the
REU file type.