Initial attempt to pass song lengths to new SID player.
Added Song length read function. The song lengths are now passed as part
of a patched BASIC rom, at $BE00. We simply use a 16K cart for this and
use some of the basic ROM that is not needed to pass our data to the
player. There are more elegant ways to do this, but this also works.