Fix display glitch in C64 and cartridge settings menu
When adding the new REU preload options, a piece of code was overlooked that
adjusts the PHI timing values at runtime for either u2plus or the u64.

This caused the option for the ultimate audio to use 16ns and 32ns as values
rather than enable/disable, and the phi timing options to use the wrong
values as well.

This commits corrects this and similar bugs in the future by looking up
CFG_C64_TIMING by id instead of index.