Implemented the UCI command 'Freeze' (Go to Ultimate menu)
In order to make it possible to go back to Ultimate menu from the SID
player, the SID player makes use of the UCI. A new command in the UCI
control target causes a button press. Some changes were required to make
this concept work, like special initialization based on specialty cart
ID, as well as an addition "no reply" bit on the UCI protocol. Also the
SID cartridge now stays selected throughout the song and only gets
deactivated when the middle button is pressed.