Fixed reset from SID player.  Release candidate for 3.2?
Added code to the UCI to reconfigure the cartridge that is selected in
the config menu. This call is now used when the SID cart detects a
reset, so that the original cartridge is loaded and started again.
TODO: Freeze button should be disabled in carts that do not support it.