Clock generation simplified to make things fit in U2.
Merge of speeddos parallel cable and U64_ulticopy fix as well. This
simplifies the next merge steps and do, or should not affect
functionality for existing builds. The U64 ulticopy fix is also related
to the clock generation, and therefore these changes are largely undone
or refactored with the current fixes.