Fixed crash of application when USB stick was removed during TAP read.
Root cause: File did not get invalidated, because the file manager
didn't know there were files open on the device. The file open function
never added the file to the open_file_list. There was no function that
ever added anything to this list... Weird!

It *should* be solid now:

When the USB stick is removed, the device gets disabled, which will
cause read errors when it is accessed without actually trying to access
the hardware anymore. USB stick is queued for removal. Then, when the
USB gets polled (why is there still polling?) the usb device gets
removed. The device gets deinstalled, and eventually the scsi driver
that is on the interface of the USB stick device, calls the file manager
to remove the root entry. The file manager then cleans up the files that
were open on that path. Files are not removed, but remain open, just
invalid. All accesses to the file manager are mutexed, so this should be
OK.

The file browser gets an invalidation event as well, which causes the
FileTypeTAP object to be removed. Either this, OR the poll of the tape
controller itself (whichever comes first) will cause the file to be
closed and cleaned up.

Seems legit, but it may be good to refactor this mechanism some day.