Made HomeDirectory work again
Also, the home directory object, with its own task, called a function
directly in the treebrowser, which is run from the UI task. This is not
correct and could have caused corruption. Now, the HomeDirectory task
simply drops a file manager event into the observer queue, which is then
handled by the ui task.