Bugfix in USB.. Improved handling of NAK.
In some situations the NAK was not handled correctly on the bus, which
could cause the whole bus to hang up. This was observed mostly with some
non-HS devices that NAK'ed in a split transaction. The solution is to
check for a timeout only on NAK, not on NYET. Timeout has now been
consistently implemented, and also the abort from the software side
(which should not be needed actually), only breaks at the right point,
and causes an interrupt with the done bit set, and status "aborted".