Rework socket DMA for disks and cartridge files
Rather than downloading the whole image into a 200KB buffer and then
writing the buffer to a ramdisk file, use a 646KB buffer and append
the data to the ramdisk file as it is received. This allows for images
larger than 200KB (eg easyflash CRT images or D81 images). With some more
work this can also be used for flashing the update and the esp32 images
using socket DMA.

Signed-off-by: Peter De Schrijver <p2@psychaos.be>