Fix incorrect keyboard shortcut actions in popups
When using keyboard shortcuts in popups an incorrect value is often returned
from poll(), causing the wrong action to be taken.

The problem is that an "active button index" is returned rather than the
expected "absolute button index".

Refactored code to return the correct value in all situations.

Bug reproduction:

1. Navigate to a random file you want to delete

2. Hit Enter

3. Select "Delete"

4. When you get the "Are you sure?" question pop, press the "Y" key on the keyboard - and you will see the file NOT being deleted

5. Try again, but press the "N" key, and the file WILL (incorrectly) be deleted