Improve GitHub workflow step names
Some of the current workflow step names could be interpreted as the step will
not include the FPGA cores. This change clarifies that the step will include
cached FPGA cores.

Updated the ESP32 build step to make it clear that if the step is skipped it is
due to the results already being in the cache.