Fix problem with alpha and beta characters in the version number
When we have alpha or beta releases the characters used for this
are actually non-printable in the ascii charset. Since JSON requires
all chars to printable, this means we are sending malformed JSON
causing browsers and other software to give errors.

This change keeps the alpha/beta characters where appropriate
and uses the words "alpha" and "beta" where ascii is required.