Improve API for CFG_TYPE_STRFUNC config settings
String settings with dynamically created lists of possible presets
(CFG_TYPE_STRFUNC) were exposed in the API just as normal strings without the
presets. This change adds a field "presets" containing the same values shown in
the Ultimate Menu. This is similar to the "values" field for CFG_TYPE_VALUE.

The reason to use "presets" and not "values" is to make clear to the caller
that an arbitrary string is also allowed (which it isn't for CFG_TYPE_VALUE
with a "values" list).